pyproject.toml
[project]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = "Homework 12"
authors = [
    {name = "serhii.zghama",email = "serhii.zghama@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "pydantic[email] (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "black (>=25.1.0,<26.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "bcrypt (<4.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "cloudinary (>=1.43.0,<2.0.0)",
    "redis (>=4.5.5,<5.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "setuptools (>=78.1.0,<79.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-cov (>=6.1.1,<7.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "sphinx (>=8.2.3,<9.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
